requirements.txt
# ============================================
# CRYPTO SIGNAL BOT - DEPENDENCIES
# ============================================
# Core Bot Dependencies
python-telegram-bot==21.4
requests==2.32.5
python-dotenv==1.0.0

# Scheduling & Async
APScheduler==3.11.1
aiohttp==3.13.2
asyncio==3.4.3

# Data Processing & Analysis
pandas==2.3.3
numpy==2.3.4

# Machine Learning
scikit-learn==1.7.2
joblib==1.5.2

# Technical Analysis
ta==0.11.0

# Charting & Visualization
matplotlib==3.10.7
mplfinance==0.12.10b0
plotly==6.4.0
seaborn==0.13.2

# Web Scraping & Parsing
beautifulsoup4==4.14.2
feedparser==6.0.12
deep-translator==1.11.4

# System Utilities
psutil==7.1.3
GitPython==3.1.45

# Security
cryptography==44.0.0

# HTTP & Networking
httpx==0.28.1
urllib3==2.5.0
